Code/gpt/gaussian-blur-project/requirements.txt
Pillow==10.0.0
numpy==1.24.3
logging==0.5.1.2